+++ R/addAQILegend.R
@@ -10,11 +10,16 @@
 #' @description This function is a convenience wrapper around
 #' \code{graphics::legend()}. It will show the AQI colors and
 #' names by default if \code{col} and \code{legend} are not specified.
+#'
+#' AQI categories are arranged with lower levels at the bottom of the legend
+#' to match the arrangement in the plot. This is different from the default
+#' "reading order" so you may wish to reverse the order of user supplied
+#' arguments with \code{rev()} .
 
 addAQILegend <- function(
   x = "topright",
   y = NULL,
-  pollutant = c("PM2.5", "CO", "OZONE"),
+  pollutant = c("PM2.5", "CO", "OZONE", "PM10"),
   palette = c("EPA", "subdued", "deuteranopia"),
   languageCode = c("eng", "spa"),
   ...
@@ -39,17 +44,18 @@ addAQILegend <- function(
   if ( ("col" %in% names(argsList)) ) {
     argsList$col <- col
   } else {
-    argsList$col <- colors
+    argsList$col <- rev(colors)      # Lower levels on the bottom
   }
 
   if ( ("legend" %in% names(argsList)) ) {
     argsList$legend <- legend
   } else {
-    argsList$legend <- names
+    argsList$legend <- rev(names)    # Lower levels on the bottom
   }
 
-  if ( !("pch" %in% names(argsList)) )
+  if ( !("pch" %in% names(argsList)) ) {
     argsList$pch <- 16
+  }
 
   if ( !("title" %in% names(argsList)) )
     argsList$title <- paste0(pollutant, " Air Quality Index")